MiniMax-M3: FP8 indexer for SM120

+++ vllm/models/minimax_m3/common/ops/index_topk.py
@@ -146,7 +146,9 @@ def _index_block_score_kernel(
             + off_k[None, :] * stride_ik_pos
             + off_d[:, None] * stride_ik_d,
         )
-        qk = tl.dot(q, k)
+        # Be explicit for the SM120 FP8 qualification path: both BF16 and E4M3
+        # inputs accumulate block scores in FP32.
+        qk = tl.dot(q, k, out_dtype=tl.float32)
         # apply causal mask as needed
         if q_start < i + BLOCK_SIZE_K:
             qk = tl.where(off_q[:, None] >= pos[None, :], qk, float("-inf"))
